requirements.txt
alpaca-py
finnhub-python
google-cloud-secret-manager
lxml
pandas
prefect>=2.0.0
prefect-slack
prefect-dask
prefect-ray
pytest
python-dotenv
requests
seaborn
ruff